Use an actor for backups
